# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-Posyandu Puskesmas Gumukmas API",
    "version": "3.7.0",
    "description": "REST API untuk autentikasi, CRUD, sinkronisasi, dan ekspor, serta endpoint GraphQL baca E-Posyandu."
  },
  "servers": [
    {
      "url": "https://e-posyandu-api.eposyandu-puskesmas-gumukmas.workers.dev"
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "summary": "Masuk memakai username dan kata sandi",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cookie sesi HttpOnly dibuat; respons tidak memuat token bearer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "summary": "Mendaftarkan faktor MFA administrator",
        "operationId": "adminMfaEnroll",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Faktor MFA dibuat dalam status belum terverifikasi" },
          "401": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/invite/complete": {
      "post": {
        "summary": "Menyelesaikan undangan administrator dan membuat sesi MFA sementara",
        "operationId": "completeAdminInvite",
        "responses": {
          "200": {
            "description": "Kata sandi disimpan dan sesi menunggu MFA",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MfaPendingLoginResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/mfa/challenge": {
      "post": {
        "summary": "Membuat challenge MFA administrator",
        "operationId": "adminMfaChallenge",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Challenge MFA dibuat" },
          "401": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "summary": "Memverifikasi TOTP atau recovery code administrator",
        "operationId": "adminMfaVerify",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Sesi administrator aktif pada AAL2 atau recovery yang valid" },
          "401": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/passkey/registration/options": {
      "post": {
        "summary": "Membuat opsi pendaftaran passkey administrator",
        "operationId": "adminPasskeyRegistrationOptions",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Opsi WebAuthn dan challenge pendaftaran" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "503": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/passkey/registration/verify": {
      "post": {
        "summary": "Memverifikasi dan menyimpan passkey administrator",
        "operationId": "adminPasskeyRegistrationVerify",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Passkey tersimpan dan sesi administrator aktif" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "422": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/passkey/authentication/options": {
      "post": {
        "summary": "Membuat opsi autentikasi passkey administrator",
        "operationId": "adminPasskeyAuthenticationOptions",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Opsi WebAuthn dan challenge autentikasi" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "503": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/passkey/authentication/verify": {
      "post": {
        "summary": "Memverifikasi passkey administrator",
        "operationId": "adminPasskeyAuthenticationVerify",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Sesi administrator aktif" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "422": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "Menghapus sesi HttpOnly dan sesi Supabase",
        "operationId": "logout",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": { "description": "Sesi dihapus" }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "summary": "Memeriksa sesi HttpOnly aktif",
        "operationId": "currentSession",
        "security": [{ "sessionCookie": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Sesi dan profil aktif" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
...
    "/api/v1/auth/presence": {
      "post": {
        "summary": "Memperbarui aktivitas akun untuk status online",
        "operationId": "touchAccountPresence",
        "security": [{ "sessionCookie": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Aktivitas akun tercatat",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PresenceResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
    "/api/v1/admin/accounts": {
      "get": {
        "summary": "Memantau seluruh akun dari halaman administrasi backend",
        "description": "Khusus super_admin dengan sesi MFA terverifikasi.",
        "operationId": "adminAccountsOverview",
        "security": [{ "sessionCookie": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Ringkasan akses penuh dan status aktivitas akun",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminAccountsOverview" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" }
        }
      },
      "post": {
        "summary": "Membuat akun dan mengirim undangan aktivasi",
        "description": "Khusus super_admin dengan sesi MFA terverifikasi.",
        "operationId": "createAdminAccount",
        "security": [{ "sessionCookie": [] }, { "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAccountInput" } } }
        },
        "responses": {
          "201": { "description": "Undangan akun dikirim" },
          "401": { "$ref": "#/components/responses/ApiError" },
          "403": { "$ref": "#/components/responses/ApiError" },
          "409": { "$ref": "#/components/responses/ApiError" },
          "422": { "$ref": "#/components/responses/ApiError" }
        }
      }
    },
...
    "/api/v1/me": {
      "get": {
        "summary": "Mengambil profil dan cakupan akses",
        "operationId": "currentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profil pengguna"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-e-posyandu-session"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "turnstileToken": {
            "type": "string"
          }
        }
      },
      "AccessProfile": {
        "type": "object",
        "required": [
          "userId",
          "role",
          "accessMode"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "desa": {
            "type": [
              "string",
              "null"
            ]
          },
          "posyandu": {
            "type": [
              "string",
              "null"
            ]
          },
          "accessMode": {
            "type": "string",
            "enum": ["read", "write"]
          }
        }
      },
      "PresenceResponse": {
        "type": "object",
        "required": ["online", "checkedAt"],
        "properties": {
          "online": { "type": "boolean", "const": true },
          "checkedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AdminAccountPresence": {
        "type": "object",
        "required": ["userId", "active", "accessMode", "isCurrentAccount", "presenceStatus"],
        "properties": {
          "userId": { "type": "string", "format": "uuid" },
          "username": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"], "format": "email" },
          "role": { "type": ["string", "null"] },
          "village": { "type": ["string", "null"] },
          "posyandu": { "type": ["string", "null"] },
          "active": { "type": "boolean" },
          "accessMode": { "type": "string", "enum": ["read", "write"] },
          "isCurrentAccount": { "type": "boolean" },
          "presenceStatus": { "type": "string", "enum": ["online", "offline"] },
          "lastSeenAt": { "type": ["string", "null"], "format": "date-time" },
          "createdAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "AdminAccountInput": {
        "type": "object",
        "required": ["email", "username", "role", "accessMode"],
        "properties": {
          "email": { "type": "string", "format": "email", "maxLength": 254 },
          "username": { "type": "string", "minLength": 3, "maxLength": 32, "pattern": "^[a-z0-9][a-z0-9._-]{2,31}$" },
          "role": { "type": "string", "enum": ["Kader Posyandu", "Bidan Desa", "Ahli Gizi", "super_admin"] },
          "village": { "type": ["string", "null"] },
          "posyandu": { "type": ["string", "null"] },
          "accessMode": { "type": "string", "enum": ["read", "write"] },
          "active": { "type": "boolean", "default": true }
        }
      },
      "AdminAccountsOverview": {
        "type": "object",
        "required": ["checkedAt", "onlineWindowSeconds", "access", "summary", "accounts"],
        "properties": {
          "checkedAt": { "type": "string", "format": "date-time" },
          "onlineWindowSeconds": { "type": "integer", "minimum": 60 },
          "access": { "type": "object" },
          "summary": {
            "type": "object",
            "required": ["total", "active", "online", "offline"],
            "properties": {
              "total": { "type": "integer", "minimum": 0 },
              "active": { "type": "integer", "minimum": 0 },
              "online": { "type": "integer", "minimum": 0 },
              "offline": { "type": "integer", "minimum": 0 }
            }
          },
          "accounts": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AdminAccountPresence" }
          }
        }
      },
...
      "AuthenticatedLoginResponse": {
        "type": "object",
        "required": [
          "user",
          "profile"
        ],
        "properties": {
          "user": {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "email"
              }
            }
          },
          "profile": {
            "$ref": "#/components/schemas/AccessProfile"
          }
        }
      },
      "MfaPendingLoginResponse": {
        "type": "object",
        "required": ["mfaRequired", "setupRequired", "factors", "expiresIn"],
        "properties": {
          "mfaRequired": { "const": true },
          "setupRequired": { "type": "boolean" },
          "expiresIn": { "type": "integer", "minimum": 1 },
          "factors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["id", "type"],
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "type": { "type": "string", "enum": ["webauthn", "totp"] },
                "name": { "type": ["string", "null"] }
              }
            }
          }
        }
      },
      "LoginResponse": {
        "oneOf": [
          { "$ref": "#/components/schemas/AuthenticatedLoginResponse" },
          { "$ref": "#/components/schemas/MfaPendingLoginResponse" }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "error": {
            "type": "object"
          }
        }
      }
    },
    "responses": {
      "ApiError": {
        "description": "Permintaan ditolak",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}